# Fix `function-linear-gradient-no-nonstandard-direction` performance