--- a/.changeset/popular-cougars-build.md
+++ b/.changeset/popular-cougars-build.md
@@ -1,0 +1,5 @@
+---
+"stylelint": patch
+---
+
+Fixed: `function-linear-gradient-no-nonstandard-direction` performance