--- a/lib/rules/function-linear-gradient-no-nonstandard-direction/index.js
+++ b/lib/rules/function-linear-gradient-no-nonstandard-direction/index.js
@@ -19,14 +19,22 @@ const meta = {
 	url: 'https://stylelint.io/user-guide/rules/function-linear-gradient-no-nonstandard-direction',
 };
 
+const LINEAR_GRADIENT_FUNCTION = /(?:-webkit-|-moz-|-o-|-ms-)?linear-gradient\(/i;
+const LINEAR_GRADIENT_FUNCTION_NAME = /^(?:-webkit-|-moz-|-o-|-ms-)?linear-gradient$/i;
+
+const DIRECTION_WITH_TO = /^to (top|left|bottom|right)(?: (top|left|bottom|right))?$/i;
+const DIRECTION_WITHOUT_TO = /^(top|left|bottom|right)(?: (top|left|bottom|right))?$/i;
+
+const LEFT_RIGHT_TOP_BOTTOM = /left|right|top|bottom/i;
+const DIGIT = /[\d.]/;
+const ANGLE = /^[\d.]+(?:deg|grad|rad|turn)$/;
+
 /**
  * @param {string} source
  * @param {boolean} withToPrefix
  */
 function isStandardDirection(source, withToPrefix) {
-	const regexp = withToPrefix
-		? /^to (top|left|bottom|right)(?: (top|left|bottom|right))?$/
-		: /^(top|left|bottom|right)(?: (top|left|bottom|right))?$/;
+	const regexp = withToPrefix ? DIRECTION_WITH_TO : DIRECTION_WITHOUT_TO;
 
 	const matches = source.match(regexp);
 
@@ -56,6 +64,8 @@ const rule = (primary) => {
 		}
 
 		root.walkDecls((decl) => {
+			if (!LINEAR_GRADIENT_FUNCTION.test(decl.value)) return;
+
 			valueParser(decl.value).walk((valueNode) => {
 				if (valueNode.type !== 'function') {
 					return;
@@ -63,7 +73,7 @@ const rule = (primary) => {
 
 				functionArgumentsSearch(
 					valueParser.stringify(valueNode).toLowerCase(),
-					/^(-webkit-|-moz-|-o-|-ms-)?linear-gradient$/i,
+					LINEAR_GRADIENT_FUNCTION_NAME,
 					(expression, expressionIndex) => {
 						const args = expression.split(',');
 						const firstArg = (args[0] || '').trim();
@@ -74,8 +84,8 @@ const rule = (primary) => {
 						}
 
 						// If the first character is a number, we can assume the user intends an angle
-						if (/[\d.]/.test(firstArg.charAt(0))) {
-							if (/^[\d.]+(?:deg|grad|rad|turn)$/.test(firstArg)) {
+						if (DIGIT.test(firstArg.charAt(0))) {
+							if (ANGLE.test(firstArg)) {
 								return;
 							}
 
@@ -87,7 +97,7 @@ const rule = (primary) => {
 						// The first argument may not be a direction: it may be an angle,
 						// or a color stop (in which case user gets default direction, "to bottom")
 						// cf. https://drafts.csswg.org/css-images-3/#linear-gradient-syntax
-						if (!/left|right|top|bottom/.test(firstArg)) {
+						if (!LEFT_RIGHT_TOP_BOTTOM.test(firstArg)) {
 							return;
 						}