--- a/lib/rules/function-linear-gradient-no-nonstandard-direction/index.js
+++ b/lib/rules/function-linear-gradient-no-nonstandard-direction/index.js
@@ -19,13 +19,14 @@ const meta = {
 	url: 'https://stylelint.io/user-guide/rules/function-linear-gradient-no-nonstandard-direction',
 };
 
-const LINEAR_GRADIENT_FUNCTION = /(?:-webkit-|-moz-|-o-|-ms-)?linear-gradient\(/i;
-const LINEAR_GRADIENT_FUNCTION_NAME = /^(?:-webkit-|-moz-|-o-|-ms-)?linear-gradient$/i;
+const LINEAR_GRADIENT_FUNCTION = '(?:-webkit-|-moz-|-o-|-ms-)?linear-gradient';
+const LINEAR_GRADIENT_FUNCTION_CALL = new RegExp(`${LINEAR_GRADIENT_FUNCTION}\\(`, 'i');
+const LINEAR_GRADIENT_FUNCTION_NAME = new RegExp(`^${LINEAR_GRADIENT_FUNCTION}$`, 'i');
 
-const DIRECTION_WITH_TO = /^to (top|left|bottom|right)(?: (top|left|bottom|right))?$/i;
-const DIRECTION_WITHOUT_TO = /^(top|left|bottom|right)(?: (top|left|bottom|right))?$/i;
+const DIRECTION = /top|left|bottom|right/i;
+const DIRECTION_WITH_TO = new RegExp(`^to (${DIRECTION.source})(?: (${DIRECTION.source}))?$`, 'i');
+const DIRECTION_WITHOUT_TO = new RegExp(`^(${DIRECTION.source})(?: (${DIRECTION.source}))?$`, 'i');
 
-const LEFT_RIGHT_TOP_BOTTOM = /left|right|top|bottom/i;
 const DIGIT = /[\d.]/;
 const ANGLE = /^[\d.]+(?:deg|grad|rad|turn)$/;
 
@@ -64,7 +65,7 @@ const rule = (primary) => {
 		}
 
 		root.walkDecls((decl) => {
-			if (!LINEAR_GRADIENT_FUNCTION.test(decl.value)) return;
+			if (!LINEAR_GRADIENT_FUNCTION_CALL.test(decl.value)) return;
 
 			valueParser(decl.value).walk((valueNode) => {
 				if (valueNode.type !== 'function') {
@@ -97,7 +98,7 @@ const rule = (primary) => {
 						// The first argument may not be a direction: it may be an angle,
 						// or a color stop (in which case user gets default direction, "to bottom")
 						// cf. https://drafts.csswg.org/css-images-3/#linear-gradient-syntax
-						if (!LEFT_RIGHT_TOP_BOTTOM.test(firstArg)) {
+						if (!DIRECTION.test(firstArg)) {
 							return;
 						}