<!-- Each pull request must be associated with an open issue unless it's a documentation fix. If a corresponding issue does not exist, please create one so we can discuss the change first. -->

<!-- Please answer the following. We close pull requests that don't. -->

> Which issue, if any, is this issue related to?

See https://github.com/stylelint/stylelint/issues/6869

> Is there anything in the PR that needs further explanation?

No, it's self-explanatory.

-----

before :

```
Warnings: 0
Mean: 89.20350494117648 ms
Deviation: 11.294021704297947 ms
```

after : 

```
Warnings: 0
Mean: 77.0127245 ms
Deviation: 7.537474958126632 ms
```